+++ tests/test_get_env.py
@@ -1,18 +1,13 @@
+import os
 from tests.mock_requests import MockResponse
 
 
-def test_info_v2_check(mocker, mock_api_v2_resource, utils_resource):
-    import afs
-
-    mocker.patch("afs.utils.urljoin")
+def test_info_v2_check():
     from afs.get_env import AfsEnv
 
-    AfsEnv()
-    afs.utils.urljoin.assert_called_once_with("http://afs.org.tw/", extra_paths={})
-
-
-def test_version_v2_check(mocker, mock_api_v2_AFS_API_VERSION_resource, utils_resource):
-    import afs
-
-    mocker.patch("afs.utils.urljoin")
-    assert mock_api_v2_AFS_API_VERSION_resource.version == "2.1.7"
+    afs_env = AfsEnv()
+    assert afs_env.afs_url == os.getenv("afs_url", None)
+    assert afs_env.instance_id == os.getenv("instance_id", None)
+    assert afs_env.target_endpoint == f"{os.getenv('afs_url', None)}/v2/"
+    assert afs_env.bucket_name != None
+    assert afs_env.version != None