+++ tests/v2/conftest.py
@@ -3,8 +3,6 @@ from tests.mock_requests import MockResponse
 from dotenv import load_dotenv
 from afs import models
 
-load_dotenv()
-
 
 @pytest.fixture(scope="session")
 def model_file():
@@ -20,6 +18,19 @@ def model_file():
 @pytest.fixture(scope="function")
 def afs_models():
     my_models = models()
+    yield my_models
+
+
+@pytest.fixture(scope="function")
+def afs_models_blob():
+    my_models = models()
+    blob_endpoint = os.getenv("blob_endpoint", None)
+    encode_blob_accessKey = os.getenv("blob_accessKey", None)
+    encode_blob_secretKey = os.getenv("blob_secretKey", None)
+
+    my_models.set_blob_credential(
+        blob_endpoint, encode_blob_accessKey, encode_blob_secretKey
+    )
     yield my_models
 
 
@@ -47,10 +58,19 @@ def delete_model_respository(afs_models):
 @pytest.fixture(scope="function")
 def delete_mr_and_model(afs_models):
     yield
-    afs_models.delete_model(
-        model_name="test_model", model_repository_name="test_model_repository"
-    )
-    afs_models.delete_model_repository(model_repository_name="test_model_repository")
+    try:
+        afs_models.delete_model(
+            model_name="test_model", model_repository_name="test_model_repository"
+        )
+    except Exception as e:
+        pass
+
+    try:
+        afs_models.delete_model_repository(
+            model_repository_name="test_model_repository"
+        )
+    except Exception as e:
+        pass
 
 
 @pytest.fixture(scope="function")
@@ -59,3 +79,38 @@ def apm_node_env():
     os.environ["PAI_DATA_DIR"] = json.dumps(pai_data_dir)
     yield
     del os.environ["PAI_DATA_DIR"]
+
+
+@pytest.fixture(scope="function")
+def error1_apm_node_env():
+    pai_data_dir = "123"
+    os.environ["PAI_DATA_DIR"] = pai_data_dir
+    yield
+    del os.environ["PAI_DATA_DIR"]
+
+
+@pytest.fixture(scope="function")
+def big_model():
+    big_model_filename = "big_model.h5"
+    if not os.path.exists(big_model_filename):
+        f = open(big_model_filename, "wb")
+        # f.seek((301 * 1024 * 1024 + 1) - 1)
+        f.seek((1 * 1024 * 1024 + 1) - 1)
+        f.write(b"\0")
+        f.close()
+
+    yield big_model_filename
+
+
+@pytest.fixture(scope="function")
+def afs_models_with_error_blob():
+    my_models = models()
+    blob_endpoint = os.getenv("blob_endpoint", None)
+    encode_blob_accessKey = os.getenv("blob_accessKey", None)
+    encode_blob_secretKey = "NDhkZTU1MGFjOTEwNGI3MTk4N2RjZGQ5ZWFjMTI0OTk="
+
+    encode_blob_accessKey = os.getenv("blob_accessKey", None)
+    my_models.set_blob_credential(
+        blob_endpoint, encode_blob_accessKey, encode_blob_secretKey
+    )
+    yield my_models