+++ tests/v2/test_models.py
@@ -1,4 +1,5 @@
 from uuid import UUID
+import pytest
 
 
 def test_create_model_repo(afs_models, delete_model_respository):
@@ -118,3 +119,73 @@ def test_create_firehose_apm_model(
         last_one=True,
     )
     assert get_resp == resp["uuid"]
+
+
+def test_error1_create_firehose_apm_model(
+    afs_models, error1_apm_node_env, delete_mr_and_model, model_file
+):
+    resp = afs_models.upload_model(
+        model_path="unit_test_model",
+        accuracy=1.0,
+        loss=1.0,
+        tags={"tag_key": "tag_value"},
+        model_repository_name="test_model_repository",
+        model_name="test_model",
+    )
+    assert isinstance(resp, dict)
+    assert "uuid" in resp
+    assert "name" in resp
+    assert "created_at" in resp
+    assert "parameters" in resp
+    assert "tags" in resp
+    assert "evaluation_result" in resp
+
+    get_resp = afs_models.get_model_id(
+        model_name="test_model",
+        model_repository_name="test_model_repository",
+        last_one=True,
+    )
+    assert get_resp == resp["uuid"]
+
+
+def test_create_big_model(afs_models_blob, big_model, delete_mr_and_model):
+    resp = afs_models_blob.upload_model(
+        model_path=big_model,
+        accuracy=1.0,
+        loss=1.0,
+        tags={"tag_key": "tag_value"},
+        model_repository_name="test_model_repository",
+        model_name="test_model",
+        blob_mode=True,
+    )
+    assert isinstance(resp, dict)
+    assert "uuid" in resp
+    assert "name" in resp
+    assert "created_at" in resp
+    assert "parameters" in resp
+    assert "tags" in resp
+    assert "evaluation_result" in resp
+    assert "size" in resp
+    assert resp["size"] > 1 * 1024 * 1024
+
+    get_resp = afs_models_blob.get_model_id(
+        model_name="test_model",
+        model_repository_name="test_model_repository",
+        last_one=True,
+    )
+    assert get_resp == resp["uuid"]
+
+
+def test_connect_blob_error_create_model(
+    afs_models_with_error_blob, big_model, delete_mr_and_model
+):
+    with pytest.raises(Exception):
+        assert afs_models_with_error_blob.upload_model(
+            model_path=big_model,
+            accuracy=1.0,
+            loss=1.0,
+            tags={"tag_key": "tag_value"},
+            model_repository_name="test_model_repository",
+            model_name="test_model",
+            blob_mode=True,
+        )